Add missing zio->io_logical assignment
The zio constructor/destructor workaround patch accidentally
introduced a regression.  One call path was missed which makes
it possible that zio->io_logicial will not get initialized for
a new zio.  This can result is the zio being mistakenly treated
as a logical IO which will result in the following assertion
when coupled with an io error.

  SPLError: 24490:0:(zio.c:2893:zio_done())
  ASSERTION(zio->io_vd == NULL && zio->io_bp != NULL) failed

Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>
Issue #602
Issue #604