Add ability to override kpkg and kdevpkg variables
This change adds the ability for a user to potentially pass in the
'require_kpkg' option and/or the 'require_kdevpkg' option to rpmbuild
and override the default values that would otherwise be assigned to kpkg
and kdevpkg. This is useful, for example, if one wants to build packages
based on a kernel other than the default (i.e. a debug kernel).

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Signed-off-by: Brian Behlendorf <behlendorf1@llnl.gov>