vdev_id: handle dm-multipath partitions
Handling partitions of dm-multipath block devices in udev rules
turns out to be less straightforward than with raw devices.  A
dm-multipath partition device shows up as a whole new dm-XX device
without a suffix identifying the partition number.  Also, the
DEVTYPE variable is set to "disk", whereas it would normally be set
to "partition" for a raw disk parition.  Therefore some ugly code is
needed in the udev helper to detect if we are handling a partition,
infer the parent device, and append a -part suffix with the
partition number.