Change zfs_txg_timeout to 1 second
The theory is our Lustre client OOM issues described in ORI-781 may be
related to the caching/commit behaviour between the Lustre client and
servers. To try and work around this, I am changing the timeout of
zfs_txg_timeout to 1 second as opposed to 5 seconds. This means the ZFS
pools will now sync at least once every second.

Signed-off-by: Prakash Surya <surya1@llnl.gov>